Fixed the cors issue in the backend as well, for all the endpoints. Did some refactoring
