+++ life-manager-rest/src/main/java/backend/app/lifemanager/features/weather/WeatherController.java
@@ -4,16 +4,16 @@ import backend.app.lifemanager.features.location.LocationService;
 import backend.app.lifemanager.features.weather.dao.locations.Location;
 import backend.app.lifemanager.features.weather.dao.weather.WeatherForecast;
 import backend.app.lifemanager.security.authentication.IAuthenticationFacade;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.core.Authentication;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
 
+@Slf4j
 @RestController
+@CrossOrigin(origins = {"http://localhost:3000", "http://localhost:9090"})
 @RequestMapping("/api/weather")
 public class WeatherController {
     private final IAuthenticationFacade authenticationFacade;
