+++ life-manager-web/src/config.tsx
@@ -1,6 +1,5 @@
 import axios from "axios";
 
 export const axiosInstance = axios.create({
-    baseURL: process.env.REACT_APP_API_BASE_URL,
-    timeout: 1000
+    baseURL: process.env.REACT_APP_API_BASE_URL
 });
